@@ -15,10 +15,11 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 2.9.0+26
+version: 3.5.1+34
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: '>=2.18.2 <3.0.0'
+  flutter: ^3.3.5
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,29 +30,30 @@ environment:
 dependencies:
   audio_service: ^0.18.7
   audio_session: ^0.1.10
-  cached_network_image: ^3.2.1
+  cached_network_image: ^3.2.2
   flutter:
     sdk: flutter
-  flutter_downloader: ^1.8.1
+  flutter_downloader: ^1.8.4
   flutter_localizations: 
     sdk: flutter
   fluttertoast: ^8.0.9
   get_it: ^7.2.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  http: ^0.13.4
+  http: ^0.13.5
   intl: ^0.17.0 
-  just_audio: ^0.9.28
+  just_audio: ^0.9.29
   material_design_icons_flutter: ^5.0.6996
   on_audio_query: ^2.6.1
-  package_info_plus: ^1.4.3
+  package_info_plus: ^3.0.1
   path_provider: ^2.0.11
-  permission_handler: ^10.0.0
-  youtube_explode_dart: ^1.12.0
+  permission_handler: ^10.2.0
+  url_launcher: ^6.1.6
+  youtube_explode_dart: ^1.12.1
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.9.3
-  flutter_native_splash: ^2.2.7
+  flutter_launcher_icons: ^0.10.0
+  flutter_native_splash: ^2.2.11
   flutter_test:
     sdk: flutter
 
@@ -81,4 +83,26 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/db/playlists.db.json
+
+
+flutter_native_splash:
+  color: "#ffffff"
+  image: assets/images/splash.png
+  color_dark: "#151515"
+  image_dark: assets/images/splash.png
+
+  android_12:
+    image: assets/images/splash.png
+    icon_background_color: "#ffffff"
+    image_dark: assets/images/splash.png
+    icon_background_color_dark: "#151515"
+
+
+flutter_icons:
+  android: "launcher_icon"
+  adaptive_icon_background: "#191919"
+  adaptive_icon_foreground: "assets/images/ic_launcher_foreground.png"
+  adaptive_icon_round: "assets/images/ic_launcher_round.png"
+  image_path: "assets/images/ic_launcher.png"
+  ios: true
 
