@@ -3,10 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:musify/API/musify.dart';
 import 'package:musify/services/audio_manager.dart';
-import 'package:musify/style/appColors.dart';
+import 'package:musify/services/download_manager.dart';
+import 'package:musify/style/appTheme.dart';
 
 class SongBar extends StatelessWidget {
-  SongBar(this.song, this.moveBackAfterPlay, {Key? key}) : super(key: key);
+  SongBar(this.song, this.moveBackAfterPlay, {super.key});
 
   late final dynamic song;
   late final bool moveBackAfterPlay;
@@ -26,26 +27,26 @@ class SongBar extends StatelessWidget {
             id = 0;
           }
           if (moveBackAfterPlay) {
-            Navigator.pop(context);
+            Navigator.pushReplacementNamed(context, '/');
           }
         },
-        splashColor: accent,
-        hoverColor: accent,
-        focusColor: accent,
-        highlightColor: accent,
+        splashColor: accent.withOpacity(0.4),
+        hoverColor: accent.withOpacity(0.4),
+        focusColor: accent.withOpacity(0.4),
+        highlightColor: accent.withOpacity(0.4),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
             CachedNetworkImage(
-              width: 70,
-              height: 70,
+              width: 60,
+              height: 60,
               imageUrl: song['lowResImage'].toString(),
               imageBuilder: (context, imageProvider) => DecoratedBox(
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(12),
                   image: DecorationImage(
                     image: imageProvider,
-                    fit: BoxFit.cover,
+                    centerSlice: const Rect.fromLTRB(1, 1, 1, 1),
                   ),
                 ),
               ),
@@ -65,9 +66,10 @@ class SongBar extends StatelessWidget {
                           .replaceAll('&quot;', '"')
                           .replaceAll('&amp;', '&'),
                       style: TextStyle(
-                          color: accent,
-                          fontSize: 16,
-                          fontWeight: FontWeight.w700),
+                        color: accent,
+                        fontSize: 16,
+                        fontWeight: FontWeight.w700,
+                      ),
                     ),
                   ),
                   const SizedBox(
@@ -78,10 +80,11 @@ class SongBar extends StatelessWidget {
                     child: Text(
                       overflow: TextOverflow.ellipsis,
                       song['more_info']['singers'].toString(),
-                      style: const TextStyle(
-                          color: Colors.white70,
-                          fontWeight: FontWeight.w400,
-                          fontSize: 14),
+                      style: TextStyle(
+                        color: Theme.of(context).hintColor,
+                        fontWeight: FontWeight.w400,
+                        fontSize: 14,
+                      ),
                     ),
                   ),
                 ],
@@ -117,7 +120,7 @@ class SongBar extends StatelessWidget {
                 IconButton(
                   color: accent,
                   icon: const Icon(MdiIcons.downloadOutline),
-                  onPressed: () => downloadSong(song),
+                  onPressed: () => downloadSong(context, song),
                 ),
               ],
             ),
