@@ -1,33 +1,65 @@
+import 'dart:async';
+
 import 'package:audio_service/audio_service.dart';
 import 'package:audio_session/audio_session.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_downloader/flutter_downloader.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:get_it/get_it.dart';
 import 'package:hive_flutter/hive_flutter.dart';
-import 'package:musify/API/musify.dart';
-import 'package:musify/helper/material_color_creator.dart';
-import 'package:musify/helper/version.dart';
 import 'package:musify/services/audio_handler.dart';
 import 'package:musify/services/audio_manager.dart';
 import 'package:musify/style/appColors.dart';
+import 'package:musify/style/appTheme.dart';
 import 'package:musify/ui/rootPage.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 GetIt getIt = GetIt.instance;
+late PackageInfo packageInfo;
+bool _interrupted = false;
+ThemeMode themeMode = ThemeMode.system;
+
+final codes = <String, String>{
+  'English': 'en',
+  'Georgian': 'ka',
+  'Chinese': 'zh',
+  'Dutch': 'nl',
+  'German': 'de',
+  'Indonesian': 'id',
+  'Italian': 'it',
+  'Polish': 'pl',
+  'Portuguese': 'pt',
+  'Spanish': 'es',
+  'Turkish': 'tr',
+  'Ukrainian': 'uk',
+};
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
-  static Future<void> setLocale(BuildContext context, Locale newLocale) async {
-    final _MyAppState state = context.findAncestorStateOfType<_MyAppState>()!;
+  static Future<void> setThemeMode(
+    BuildContext context,
+    ThemeMode newThemeMode,
+  ) async {
+    final state = context.findAncestorStateOfType<_MyAppState>()!;
+    state.changeTheme(newThemeMode);
+  }
+
+  static Future<void> setLocale(
+    BuildContext context,
+    Locale newLocale,
+  ) async {
+    final state = context.findAncestorStateOfType<_MyAppState>()!;
     state.changeLanguage(newLocale);
   }
 
   static Future<void> setAccentColor(
-      BuildContext context, Color newAccentColor) async {
-    final _MyAppState state = context.findAncestorStateOfType<_MyAppState>()!;
+    BuildContext context,
+    Color newAccentColor,
+  ) async {
+    final state = context.findAncestorStateOfType<_MyAppState>()!;
     state.changeAccentColor(newAccentColor);
   }
 
@@ -38,6 +70,12 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   Locale _locale = const Locale('en', '');
 
+  void changeTheme(ThemeMode newThemeMode) {
+    setState(() {
+      themeMode = newThemeMode;
+    });
+  }
+
   void changeLanguage(Locale locale) {
     setState(() {
       _locale = locale;
@@ -46,74 +84,38 @@ class _MyAppState extends State<MyApp> {
 
   void changeAccentColor(Color newAccentColor) {
     setState(() {
-      accent = newAccentColor;
+      accent = getMaterialColorFromColor(newAccentColor);
     });
   }
 
   @override
   void initState() {
     super.initState();
-    getLocalSongs();
-    final Map<String, String> codes = {
-      'English': 'en',
-      'Georgian': 'ka',
-      'Chinese': 'zh',
-      'Dutch': 'nl',
-      'German': 'de',
-      'Indonesian': 'id',
-      'Italian': 'it',
-      'Polish': 'pl',
-      'Portuguese': 'pt',
-      'Spanish': 'es',
-      'Turkish': 'tr',
-      'Ukrainian': 'uk',
-    };
-    _locale = Locale(codes[Hive.box('settings')
-        .get('language', defaultValue: 'English') as String]!);
+    _locale = Locale(
+      codes[Hive.box('settings').get('language', defaultValue: 'English')
+          as String]!,
+    );
+    themeMode = Hive.box('settings').get('themeMode', defaultValue: 'system') ==
+            'system'
+        ? ThemeMode.system
+        : Hive.box('settings').get('themeMode') == 'light'
+            ? ThemeMode.light
+            : ThemeMode.dark;
   }
 
   @override
   void dispose() {
     Hive.close();
-    audioPlayer!.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      themeMode: themeMode,
       debugShowCheckedModeBanner: false,
-      themeMode: ThemeMode.dark,
-      darkTheme: ThemeData(
-        scaffoldBackgroundColor: bgColor,
-        canvasColor: bgColor,
-        appBarTheme: AppBarTheme(backgroundColor: bgColor),
-        colorScheme:
-            ColorScheme.fromSwatch(primarySwatch: createMaterialColor(accent)),
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        fontFamily: 'Ubuntu',
-        useMaterial3: true,
-        pageTransitionsTheme: const PageTransitionsTheme(
-          builders: {
-            TargetPlatform.android: ZoomPageTransitionsBuilder(),
-          },
-        ),
-      ),
-      theme: ThemeData(
-        scaffoldBackgroundColor: Colors.white,
-        canvasColor: Colors.white,
-        appBarTheme: const AppBarTheme(backgroundColor: Colors.white),
-        colorScheme:
-            ColorScheme.fromSwatch(primarySwatch: createMaterialColor(accent)),
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        fontFamily: 'Ubuntu',
-        useMaterial3: true,
-        pageTransitionsTheme: const PageTransitionsTheme(
-          builders: {
-            TargetPlatform.android: ZoomPageTransitionsBuilder(),
-          },
-        ),
-      ),
+      darkTheme: getAppDarkTheme(),
+      theme: getAppLightTheme(),
       localizationsDelegates: const [
         AppLocalizations.delegate,
         GlobalMaterialLocalizations.delegate,
@@ -141,7 +143,10 @@ class _MyAppState extends State<MyApp> {
         Locale('ur', '')
       ],
       locale: _locale,
-      home: Musify(),
+      initialRoute: '/',
+      routes: {
+        '/': (context) => Musify(),
+      },
     );
   }
 }
@@ -152,17 +157,6 @@ void main() async {
   await Hive.openBox('settings');
   await Hive.openBox('user');
   await Hive.openBox('cache');
-  await FlutterDownloader.initialize(
-    debug:
-        true, // optional: set to false to disable printing logs to console (default: true)
-    ignoreSsl:
-        true // option: set to false to disable working with http links (default: false)
-    ,
-  );
-  FlutterDownloader.registerCallback(TestClass.callback);
-  final PackageInfo packageInfo = await PackageInfo.fromPlatform();
-  version = packageInfo.version;
-  await enableBooster();
   await initialisation();
   runApp(const MyApp());
 }
@@ -170,7 +164,27 @@ void main() async {
 Future<void> initialisation() async {
   final session = await AudioSession.instance;
   await session.configure(const AudioSessionConfiguration.music());
-  final AudioHandler audioHandler = await AudioService.init(
+  session.interruptionEventStream.listen((event) {
+    if (event.begin) {
+      if (audioPlayer.playing) {
+        pause();
+        _interrupted = true;
+      }
+    } else {
+      switch (event.type) {
+        case AudioInterruptionType.pause:
+        case AudioInterruptionType.duck:
+          if (!audioPlayer.playing && _interrupted) {
+            play();
+          }
+          break;
+        case AudioInterruptionType.unknown:
+          break;
+      }
+      _interrupted = false;
+    }
+  });
+  final audioHandler = await AudioService.init(
     builder: MyAudioHandler.new,
     config: const AudioServiceConfig(
       androidNotificationChannelId: 'com.gokadzev.musify',
@@ -181,9 +195,23 @@ Future<void> initialisation() async {
     ),
   );
   getIt.registerSingleton<AudioHandler>(audioHandler);
+  await enableBooster();
+
+  packageInfo = await PackageInfo.fromPlatform();
+
+  try {
+    await FlutterDownloader.initialize(
+      debug: kDebugMode,
+      ignoreSsl: true,
+    );
+
+    await FlutterDownloader.registerCallback(downloadCallback);
+  } catch (e) {
+    if (kDebugMode) {
+      print(e);
+    }
+  }
 }
 
-// ignore: avoid_classes_with_only_static_members
-class TestClass {
-  static void callback(String id, DownloadTaskStatus status, int progress) {}
-}
+@pragma('vm:entry-point')
+void downloadCallback(String id, DownloadTaskStatus status, int progress) {}
