@@ -3,15 +3,16 @@ import 'dart:math';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:fluttertoast/fluttertoast.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:musify/API/musify.dart';
 import 'package:musify/customWidgets/song_bar.dart';
 import 'package:musify/customWidgets/spinner.dart';
+import 'package:musify/helper/flutter_toast.dart';
 import 'package:musify/style/appColors.dart';
+import 'package:musify/style/appTheme.dart';
 
 class PlaylistPage extends StatefulWidget {
-  const PlaylistPage({Key? key, required this.playlist}) : super(key: key);
+  const PlaylistPage({super.key, required this.playlist});
   final dynamic playlist;
 
   @override
@@ -23,7 +24,7 @@ class _PlaylistPageState extends State<PlaylistPage> {
 
   bool _isLoading = true;
   bool _hasMore = true;
-  final _itemsPerPage = 10;
+  final _itemsPerPage = 35;
   var _currentPage = 0;
   var _currentLastLoadedId = 0;
 
@@ -60,10 +61,10 @@ class _PlaylistPageState extends State<PlaylistPage> {
 
   Future<List> fetch() async {
     final list = [];
-    final int _count = widget.playlist['list'].length as int;
+    final _count = widget.playlist['list'].length as int;
     final n = min(_itemsPerPage, _count - _currentPage * _itemsPerPage);
     await Future.delayed(const Duration(seconds: 1), () {
-      for (int i = 0; i < n; i++) {
+      for (var i = 0; i < n; i++) {
         list.add(widget.playlist['list'][_currentLastLoadedId]);
         _currentLastLoadedId++;
       }
@@ -103,9 +104,6 @@ class _PlaylistPageState extends State<PlaylistPage> {
                     height: 250,
                     width: 250,
                     child: Card(
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(8),
-                      ),
                       color: Colors.transparent,
                       child: widget.playlist['image'] != ''
                           ? DecoratedBox(
@@ -124,12 +122,7 @@ class _PlaylistPageState extends State<PlaylistPage> {
                               height: 200,
                               decoration: BoxDecoration(
                                 borderRadius: BorderRadius.circular(10),
-                                gradient: const LinearGradient(
-                                  colors: [
-                                    Color.fromARGB(30, 255, 255, 255),
-                                    Color.fromARGB(30, 233, 233, 233),
-                                  ],
-                                ),
+                                color: const Color.fromARGB(30, 255, 255, 255),
                               ),
                               child: Column(
                                 mainAxisAlignment: MainAxisAlignment.center,
@@ -178,29 +171,20 @@ class _PlaylistPageState extends State<PlaylistPage> {
                           setActivePlaylist(
                             widget.playlist['list'] as List,
                           ),
-                          Fluttertoast.showToast(
-                            msg: AppLocalizations.of(context)!.queueInitText,
-                            toastLength: Toast.LENGTH_LONG,
-                            gravity: ToastGravity.BOTTOM,
-                            backgroundColor: accent,
-                            textColor: accent != const Color(0xFFFFFFFF)
-                                ? Colors.white
-                                : Colors.black,
+                          showToast(
+                            AppLocalizations.of(context)!.queueInitText,
                           ),
-                          Navigator.pop(context, false)
+                          Navigator.pushReplacementNamed(context, '/'),
                         },
-                        style: ElevatedButton.styleFrom(
-                          primary: accent,
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.circular(8),
-                          ),
+                        style: ButtonStyle(
+                          backgroundColor:
+                              MaterialStateProperty.all<Color>(accent),
                         ),
                         child: Text(
                           AppLocalizations.of(context)!.playAll.toUpperCase(),
                           style: TextStyle(
-                              color: accent != const Color(0xFFFFFFFF)
-                                  ? Colors.white
-                                  : Colors.black),
+                            color: isAccentWhite(),
+                          ),
                         ),
                       ),
                     ],
