@@ -1,6 +1,19 @@
-package com.gokadzev.musify
-
+import android.os.Build
+import android.os.Bundle
+import androidx.core.view.WindowCompat
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
+class MainActivity : FlutterActivity() {
+  override fun onCreate(savedInstanceState: Bundle?) {
+    // Aligns the Flutter view vertically with the window.
+    WindowCompat.setDecorFitsSystemWindows(getWindow(), false)
+
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+      // Disable the Android splash screen fade out animation to avoid
+      // a flicker before the similar frame is drawn in Flutter.
+      splashScreen.setOnExitAnimationListener { splashScreenView -> splashScreenView.remove() }
+    }
+
+    super.onCreate(savedInstanceState)
+  }
 }
