@@ -1,10 +1,22 @@
-// cr: https://pub.dev/packages/delayed_display
+// pub: https://pub.dev/packages/delayed_display
+// license: https://raw.githubusercontent.com/ThomasEcalle/delayed_display/master/LICENSE
+// remade (not original)
 
 import 'dart:async';
 
 import 'package:flutter/material.dart';
 
 class DelayedDisplay extends StatefulWidget {
+  /// DelayedDisplay constructor
+  const DelayedDisplay({
+    required this.child,
+    this.delay = Duration.zero,
+    this.fadingDuration = const Duration(milliseconds: 800),
+    this.slidingCurve = Curves.decelerate,
+    this.slidingBeginOffset = const Offset(0, 0.35),
+    this.fadeIn = true,
+  });
+
   /// Child that will be displayed with the animation and delay
   final Widget child;
 
@@ -22,16 +34,6 @@ class DelayedDisplay extends StatefulWidget {
 
   /// If true, make the child appear, disappear otherwise. Default to true.
   final bool fadeIn;
-
-  /// DelayedDisplay constructor
-  const DelayedDisplay({
-    required this.child,
-    this.delay = Duration.zero,
-    this.fadingDuration = const Duration(milliseconds: 800),
-    this.slidingCurve = Curves.decelerate,
-    this.slidingBeginOffset = const Offset(0, 0.35),
-    this.fadeIn = true,
-  });
 
   @override
   _DelayedDisplayState createState() => _DelayedDisplayState();
@@ -74,7 +76,7 @@ class _DelayedDisplayState extends State<DelayedDisplay>
       duration: opacityTransitionDuration,
     );
 
-    final CurvedAnimation curvedAnimation = CurvedAnimation(
+    final curvedAnimation = CurvedAnimation(
       curve: slidingCurve,
       parent: _opacityController,
     );
