@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:audio_service/audio_service.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
@@ -10,20 +8,9 @@ import 'package:musify/API/musify.dart';
 import 'package:musify/customWidgets/spinner.dart';
 import 'package:musify/helper/mediaitem.dart';
 import 'package:musify/services/audio_manager.dart';
-import 'package:musify/style/appColors.dart';
+import 'package:musify/services/download_manager.dart';
+import 'package:musify/style/appTheme.dart';
 import 'package:on_audio_query/on_audio_query.dart';
-
-String status = 'hidden';
-
-typedef OnError = void Function(Exception exception);
-
-StreamSubscription? positionSubscription;
-StreamSubscription? durationSubscription;
-
-Duration? duration;
-Duration? position;
-
-enum MPlayerState { stopped, playing, paused, loading }
 
 class AudioApp extends StatefulWidget {
   @override
@@ -33,43 +20,28 @@ class AudioApp extends StatefulWidget {
 @override
 class AudioAppState extends State<AudioApp> {
   @override
-  void initState() {
-    super.initState();
-
-    positionSubscription = audioPlayer?.positionStream
-        .listen((p) => {if (mounted) setState(() => position = p)});
-    durationSubscription = audioPlayer?.durationStream.listen(
-      (d) => {
-        if (mounted) {setState(() => duration = d)}
-      },
-    );
-  }
-
-  @override
-  void dispose() {
-    positionSubscription!.cancel();
-    durationSubscription!.cancel();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
     final size = MediaQuery.of(context).size;
     return Scaffold(
       appBar: AppBar(
         elevation: 0,
         centerTitle: true,
+        toolbarHeight: size.height * 0.09,
         title: Text(
           AppLocalizations.of(context)!.nowPlaying,
           style: TextStyle(
             color: accent,
-            fontSize: 25,
+            fontSize: size.height * 0.036,
             fontWeight: FontWeight.w700,
           ),
         ),
         leading: Padding(
           padding: const EdgeInsets.only(left: 14),
           child: IconButton(
+            focusColor: Colors.transparent,
+            splashColor: Colors.transparent,
+            hoverColor: Colors.transparent,
+            highlightColor: Colors.transparent,
             icon: Icon(
               Icons.keyboard_arrow_down,
               size: 32,
@@ -80,164 +52,131 @@ class AudioAppState extends State<AudioApp> {
         ),
       ),
       body: SingleChildScrollView(
-        child: Padding(
-          padding: EdgeInsets.only(top: size.height * 0.012),
-          child: StreamBuilder<SequenceState?>(
-            stream: audioPlayer!.sequenceStateStream,
-            builder: (context, snapshot) {
-              final state = snapshot.data;
-              if (state?.sequence.isEmpty ?? true) {
-                return const SizedBox();
-              }
-              final metadata = state!.currentSource!.tag;
-              final songLikeStatus = ValueNotifier<bool>(
-                isSongAlreadyLiked(metadata.extras['ytid']),
-              );
-              return Column(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  if (metadata.extras['localSongId'] is int)
-                    QueryArtworkWidget(
-                      id: metadata.extras['localSongId'] as int,
-                      type: ArtworkType.AUDIO,
-                      artworkBorder: BorderRadius.circular(8),
-                      artworkQuality: FilterQuality.high,
-                      quality: 100,
-                      artworkWidth: size.width / 1.2,
-                      artworkHeight: size.width / 1.2,
-                      nullArtworkWidget: SizedBox(
-                        width: size.width / 1.2,
-                        height: size.width / 1.2,
-                        child: CachedNetworkImage(
-                          imageUrl: metadata.artUri.toString(),
-                          imageBuilder: (context, imageProvider) =>
-                              DecoratedBox(
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(10),
-                              image: DecorationImage(
-                                image: imageProvider,
-                                fit: BoxFit.cover,
-                              ),
-                            ),
+        child: StreamBuilder<SequenceState?>(
+          stream: audioPlayer.sequenceStateStream,
+          builder: (context, snapshot) {
+            final state = snapshot.data;
+            if (state?.sequence.isEmpty ?? true) {
+              return const SizedBox();
+            }
+            final metadata = state!.currentSource!.tag;
+            final songLikeStatus = ValueNotifier<bool>(
+              isSongAlreadyLiked(metadata.extras['ytid']),
+            );
+            return Column(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                if (metadata.extras['localSongId'] is int)
+                  QueryArtworkWidget(
+                    id: metadata.extras['localSongId'] as int,
+                    type: ArtworkType.AUDIO,
+                    artworkBorder: BorderRadius.circular(8),
+                    artworkQuality: FilterQuality.high,
+                    quality: 100,
+                    artworkWidth: size.width / 1.2,
+                    artworkHeight: size.height / 2.7,
+                    nullArtworkWidget: Container(
+                      width: size.width / 1.2,
+                      height: size.height / 2.7,
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(10),
+                        color: const Color.fromARGB(30, 255, 255, 255),
+                      ),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: <Widget>[
+                          Icon(
+                            MdiIcons.musicNoteOutline,
+                            size: size.width / 8,
+                            color: accent,
                           ),
-                          placeholder: (context, url) => const Spinner(),
-                          errorWidget: (context, url, error) => Container(
-                            width: size.width / 1.2,
-                            height: size.width / 1.2,
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(10),
-                              gradient: const LinearGradient(
-                                colors: [
-                                  Color.fromARGB(30, 255, 255, 255),
-                                  Color.fromARGB(30, 233, 233, 233),
-                                ],
-                              ),
-                            ),
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: <Widget>[
-                                Icon(
-                                  MdiIcons.musicNoteOutline,
-                                  size: size.width / 8,
-                                  color: accent,
-                                ),
-                              ],
-                            ),
+                        ],
+                      ),
+                    ),
+                    keepOldArtwork: true,
+                  )
+                else
+                  SizedBox(
+                    width: size.width / 1.2,
+                    height: size.height / 2.7,
+                    child: CachedNetworkImage(
+                      imageUrl: metadata.artUri.toString(),
+                      imageBuilder: (context, imageProvider) => DecoratedBox(
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10),
+                          image: DecorationImage(
+                            image: imageProvider,
+                            fit: BoxFit.cover,
                           ),
                         ),
                       ),
-                      keepOldArtwork: true,
-                    )
-                  else
-                    SizedBox(
-                      width: size.width / 1.2,
-                      height: size.width / 1.2,
-                      child: CachedNetworkImage(
-                        imageUrl: metadata.artUri.toString(),
-                        imageBuilder: (context, imageProvider) => DecoratedBox(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10),
-                            image: DecorationImage(
-                              image: imageProvider,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
+                      placeholder: (context, url) => const Spinner(),
+                      errorWidget: (context, url, error) => DecoratedBox(
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10),
+                          color: const Color.fromARGB(30, 255, 255, 255),
                         ),
-                        placeholder: (context, url) => const Spinner(),
-                        errorWidget: (context, url, error) => Container(
-                          width: size.width / 1.2,
-                          height: size.width / 1.2,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10),
-                            gradient: const LinearGradient(
-                              colors: [
-                                Color.fromARGB(30, 255, 255, 255),
-                                Color.fromARGB(30, 233, 233, 233),
-                              ],
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: <Widget>[
+                            Icon(
+                              MdiIcons.musicNoteOutline,
+                              size: size.width / 8,
+                              color: accent,
                             ),
-                          ),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: <Widget>[
-                              Icon(
-                                MdiIcons.musicNoteOutline,
-                                size: size.width / 8,
-                                color: accent,
-                              ),
-                            ],
-                          ),
+                          ],
                         ),
                       ),
                     ),
-                  Padding(
-                    padding: EdgeInsets.only(
-                        top: size.height * 0.04, bottom: size.height * 0.01),
-                    child: Column(
-                      children: <Widget>[
-                        Text(
-                          metadata!.title
-                              .toString()
-                              .split(' (')[0]
-                              .split('|')[0]
-                              .trim(),
-                          textScaleFactor: 2.5,
+                  ),
+                Padding(
+                  padding: EdgeInsets.only(
+                    top: size.height * 0.04,
+                    bottom: size.height * 0.01,
+                  ),
+                  child: Column(
+                    children: <Widget>[
+                      Text(
+                        metadata!.title
+                            .toString()
+                            .split(' (')[0]
+                            .split('|')[0]
+                            .trim(),
+                        maxLines: 1,
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          fontSize: size.height * 0.035,
+                          fontWeight: FontWeight.w700,
+                          color: accent,
+                        ),
+                      ),
+                      Padding(
+                        padding: const EdgeInsets.only(top: 8),
+                        child: Text(
+                          '${metadata!.artist}',
                           maxLines: 1,
                           textAlign: TextAlign.center,
                           style: TextStyle(
-                            fontSize: 12,
-                            fontWeight: FontWeight.w700,
-                            color: accent,
+                            fontSize: size.height * 0.015,
+                            fontWeight: FontWeight.w500,
                           ),
                         ),
-                        Padding(
-                          padding: const EdgeInsets.only(top: 8),
-                          child: Text(
-                            '${metadata!.artist}',
-                            maxLines: 1,
-                            textAlign: TextAlign.center,
-                            style: TextStyle(
-                              color: accentLight,
-                              fontSize: 15,
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
+                      ),
+                    ],
                   ),
-                  Material(
-                    child: _buildPlayer(
-                      size,
-                      songLikeStatus,
-                      metadata.extras['ytid'],
-                      metadata,
-                    ),
+                ),
+                Material(
+                  child: _buildPlayer(
+                    size,
+                    songLikeStatus,
+                    metadata.extras['ytid'],
+                    metadata,
                   ),
-                ],
-              );
-            },
-          ),
+                ),
+              ],
+            );
+          },
         ),
       ),
     );
@@ -260,24 +199,62 @@ class AudioAppState extends State<AudioApp> {
           mainAxisSize: MainAxisSize.min,
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            if (duration != null)
-              Slider(
-                activeColor: accent,
-                inactiveColor: Colors.green[50],
-                value: position?.inMilliseconds.toDouble() ?? 0.0,
-                onChanged: (double? value) {
-                  setState(() {
-                    audioPlayer!.seek(
-                      Duration(
-                        milliseconds: value!.round(),
-                      ),
+            ValueListenableBuilder<Duration?>(
+              valueListenable: position,
+              builder: (_, positionvalue, __) {
+                return ValueListenableBuilder<Duration?>(
+                  valueListenable: duration,
+                  builder: (_, durationvalue, __) {
+                    return Column(
+                      mainAxisSize: MainAxisSize.min,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        if (durationvalue != null)
+                          Slider(
+                            activeColor: accent,
+                            inactiveColor: Colors.green[50],
+                            value:
+                                positionvalue?.inMilliseconds.toDouble() ?? 0.0,
+                            onChanged: (double? value) {
+                              setState(() {
+                                audioPlayer.seek(
+                                  Duration(
+                                    milliseconds: value!.round(),
+                                  ),
+                                );
+                                value = value;
+                              });
+                            },
+                            max: durationvalue.inMilliseconds.toDouble(),
+                          ),
+                        Row(
+                          mainAxisSize: MainAxisSize.max,
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            if (positionvalue != null)
+                              Text(
+                                '$positionText '.replaceFirst('0:0', '0'),
+                                style: TextStyle(
+                                  fontSize: 18,
+                                  color: Theme.of(context).hintColor,
+                                ),
+                              ),
+                            if (positionvalue != null)
+                              Text(
+                                durationText.replaceAll('0:', ''),
+                                style: TextStyle(
+                                  fontSize: 18,
+                                  color: Theme.of(context).hintColor,
+                                ),
+                              )
+                          ],
+                        )
+                      ],
                     );
-                    value = value;
-                  });
-                },
-                max: duration!.inMilliseconds.toDouble(),
-              ),
-            if (position != null) _buildProgressView(),
+                  },
+                );
+              },
+            ),
             Padding(
               padding: EdgeInsets.only(top: size.height * 0.03),
               child: Column(
@@ -288,26 +265,45 @@ class AudioAppState extends State<AudioApp> {
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: <Widget>[
                         if (metadata.extras['ytid'].toString().isNotEmpty)
-                          IconButton(
-                            padding: EdgeInsets.zero,
-                            icon: const Icon(
-                              MdiIcons.download,
-                              color: Colors.white,
-                            ),
-                            iconSize: size.width * 0.056,
-                            splashColor: Colors.transparent,
-                            onPressed: () {
-                              downloadSong(
-                                mediaItemToMap(metadata as MediaItem),
-                              );
-                            },
+                          Column(
+                            children: [
+                              IconButton(
+                                padding: EdgeInsets.zero,
+                                icon: Icon(
+                                  MdiIcons.download,
+                                  color: Theme.of(context).hintColor,
+                                ),
+                                iconSize: size.width * 0.056,
+                                splashColor: Colors.transparent,
+                                onPressed: () {
+                                  downloadSong(
+                                    context,
+                                    mediaItemToMap(metadata as MediaItem),
+                                  );
+                                },
+                              ),
+                              IconButton(
+                                padding: EdgeInsets.zero,
+                                icon: Icon(
+                                  sponsorBlockSupport.value
+                                      ? MdiIcons.playCircle
+                                      : MdiIcons.playCircleOutline,
+                                  color: Theme.of(context).hintColor,
+                                ),
+                                iconSize: size.width * 0.056,
+                                splashColor: Colors.transparent,
+                                onPressed: () =>
+                                    setState(changeSponsorBlockStatus),
+                              ),
+                            ],
                           ),
                         IconButton(
                           padding: EdgeInsets.zero,
                           icon: Icon(
                             MdiIcons.shuffle,
-                            color:
-                                shuffleNotifier.value ? accent : Colors.white,
+                            color: shuffleNotifier.value
+                                ? accent
+                                : Theme.of(context).hintColor,
                           ),
                           iconSize: size.width * 0.056,
                           onPressed: changeShuffleStatus,
@@ -317,7 +313,9 @@ class AudioAppState extends State<AudioApp> {
                           padding: EdgeInsets.zero,
                           icon: Icon(
                             Icons.skip_previous,
-                            color: hasPrevious ? Colors.white : Colors.grey,
+                            color: hasPrevious
+                                ? Theme.of(context).hintColor
+                                : Colors.grey,
                             size: size.width * 0.1,
                           ),
                           iconSize: size.width * 0.056,
@@ -329,20 +327,54 @@ class AudioAppState extends State<AudioApp> {
                             color: accent,
                             borderRadius: BorderRadius.circular(100),
                           ),
-                          child: StreamBuilder<PlayerState>(
-                            stream: audioPlayer!.playerStateStream,
-                            builder: (context, snapshot) {
-                              if (snapshot.hasData) {
-                                final playerState = snapshot.data;
-                                return _playerControllers(playerState!, size);
-                              } else {
+                          child: ValueListenableBuilder<PlayerState>(
+                            valueListenable: playerState,
+                            builder: (_, value, __) {
+                              if (value.processingState ==
+                                      ProcessingState.loading ||
+                                  value.processingState ==
+                                      ProcessingState.buffering) {
                                 return Container(
                                   margin: const EdgeInsets.all(8),
                                   width: size.width * 0.08,
                                   height: size.width * 0.08,
-                                  child: const CircularProgressIndicator(
+                                  child: CircularProgressIndicator(
                                     valueColor: AlwaysStoppedAnimation<Color>(
-                                        Color.fromARGB(255, 0, 0, 0)),
+                                      Theme.of(context).hintColor,
+                                    ),
+                                  ),
+                                );
+                              } else if (value.playing != true) {
+                                return IconButton(
+                                  icon: Icon(
+                                    MdiIcons.play,
+                                    color: Theme.of(context).hintColor,
+                                  ),
+                                  iconSize: size.width * 0.1,
+                                  onPressed: play,
+                                  splashColor: Colors.transparent,
+                                );
+                              } else if (value.processingState !=
+                                  ProcessingState.completed) {
+                                return IconButton(
+                                  icon: Icon(
+                                    MdiIcons.pause,
+                                    color: Theme.of(context).hintColor,
+                                  ),
+                                  iconSize: size.width * 0.1,
+                                  onPressed: pause,
+                                  splashColor: Colors.transparent,
+                                );
+                              } else {
+                                return IconButton(
+                                  icon: Icon(
+                                    MdiIcons.replay,
+                                    color: Theme.of(context).hintColor,
+                                  ),
+                                  iconSize: size.width * 0.056,
+                                  onPressed: () => audioPlayer.seek(
+                                    Duration.zero,
+                                    index: audioPlayer.effectiveIndices!.first,
                                   ),
                                 );
                               }
@@ -353,7 +385,9 @@ class AudioAppState extends State<AudioApp> {
                           padding: EdgeInsets.zero,
                           icon: Icon(
                             Icons.skip_next,
-                            color: hasNext ? Colors.white : Colors.grey,
+                            color: hasNext
+                                ? Theme.of(context).hintColor
+                                : Colors.grey,
                             size: size.width * 0.1,
                           ),
                           iconSize: size.width * 0.08,
@@ -364,7 +398,9 @@ class AudioAppState extends State<AudioApp> {
                           padding: EdgeInsets.zero,
                           icon: Icon(
                             MdiIcons.repeat,
-                            color: repeatNotifier.value ? accent : Colors.white,
+                            color: repeatNotifier.value
+                                ? accent
+                                : Theme.of(context).hintColor,
                           ),
                           iconSize: size.width * 0.056,
                           onPressed: changeLoopStatus,
@@ -373,21 +409,6 @@ class AudioAppState extends State<AudioApp> {
                         if (metadata.extras['ytid'].toString().isNotEmpty)
                           Column(
                             children: [
-                              ValueListenableBuilder<bool>(
-                                valueListenable: playNextSongAutomatically,
-                                builder: (_, value, __) {
-                                  return IconButton(
-                                    padding: EdgeInsets.zero,
-                                    icon: Icon(
-                                      MdiIcons.chevronRight,
-                                      color: value ? accent : Colors.white,
-                                    ),
-                                    iconSize: size.width * 0.056,
-                                    splashColor: Colors.transparent,
-                                    onPressed: changeAutoPlayNextStatus,
-                                  );
-                                },
-                              ),
                               ValueListenableBuilder<bool>(
                                 valueListenable: songLikeStatus,
                                 builder: (_, value, __) {
@@ -404,7 +425,7 @@ class AudioAppState extends State<AudioApp> {
                                     );
                                   } else {
                                     return IconButton(
-                                      color: Colors.white,
+                                      color: Theme.of(context).hintColor,
                                       icon: const Icon(MdiIcons.starOutline),
                                       iconSize: size.width * 0.056,
                                       splashColor: Colors.transparent,
@@ -414,6 +435,25 @@ class AudioAppState extends State<AudioApp> {
                                       },
                                     );
                                   }
+                                },
+                              ),
+                              ValueListenableBuilder<bool>(
+                                valueListenable: playNextSongAutomatically,
+                                builder: (_, value, __) {
+                                  return IconButton(
+                                    padding: EdgeInsets.zero,
+                                    icon: Icon(
+                                      value
+                                          ? MdiIcons.skipNextCircle
+                                          : MdiIcons.skipNextCircleOutline,
+                                      color: value
+                                          ? accent
+                                          : Theme.of(context).hintColor,
+                                    ),
+                                    iconSize: size.width * 0.056,
+                                    splashColor: Colors.transparent,
+                                    onPressed: changeAutoPlayNextStatus,
+                                  );
                                 },
                               ),
                             ],
@@ -437,7 +477,6 @@ class AudioAppState extends State<AudioApp> {
                                 context: context,
                                 builder: (context) => Container(
                                   decoration: const BoxDecoration(
-                                    color: Color(0xFF151515),
                                     borderRadius: BorderRadius.only(
                                       topLeft: Radius.circular(18),
                                       topRight: Radius.circular(18),
@@ -498,9 +537,8 @@ class AudioAppState extends State<AudioApp> {
                                                   child: SingleChildScrollView(
                                                     child: Text(
                                                       value,
-                                                      style: TextStyle(
+                                                      style: const TextStyle(
                                                         fontSize: 16,
-                                                        color: accentLight,
                                                       ),
                                                       textAlign:
                                                           TextAlign.center,
@@ -511,7 +549,8 @@ class AudioAppState extends State<AudioApp> {
                                             );
                                           } else if (value == 'null') {
                                             return const SizedBox(
-                                                child: Spinner());
+                                              child: Spinner(),
+                                            );
                                           } else {
                                             return Padding(
                                               padding: const EdgeInsets.only(
@@ -523,8 +562,7 @@ class AudioAppState extends State<AudioApp> {
                                                     context,
                                                   )!
                                                       .lyricsNotAvailable,
-                                                  style: TextStyle(
-                                                    color: accentLight,
+                                                  style: const TextStyle(
                                                     fontSize: 25,
                                                   ),
                                                 ),
@@ -552,64 +590,4 @@ class AudioAppState extends State<AudioApp> {
           ],
         ),
       );
-
-  Row _buildProgressView() => Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Text(
-            position != null
-                ? '$positionText '.replaceFirst('0:0', '0')
-                : duration != null
-                    ? durationText
-                    : '',
-            style: const TextStyle(fontSize: 18, color: Colors.white),
-          ),
-          const Spacer(),
-          Text(
-            position != null
-                ? durationText.replaceAll('0:', '')
-                : duration != null
-                    ? durationText
-                    : '',
-            style: const TextStyle(fontSize: 18, color: Colors.white),
-          )
-        ],
-      );
-
-  Widget _playerControllers(PlayerState playerState, Size size) {
-    final processingState = playerState.processingState;
-    if (processingState == ProcessingState.loading ||
-        processingState == ProcessingState.buffering) {
-      return Container(
-        margin: const EdgeInsets.all(8),
-        width: size.width * 0.08,
-        height: size.width * 0.08,
-        child: const CircularProgressIndicator(
-          valueColor:
-              AlwaysStoppedAnimation<Color>(Color.fromARGB(255, 0, 0, 0)),
-        ),
-      );
-    } else if (audioPlayer!.playing != true) {
-      return IconButton(
-        icon: const Icon(MdiIcons.play),
-        iconSize: size.width * 0.1,
-        onPressed: play,
-        splashColor: Colors.transparent,
-      );
-    } else if (processingState != ProcessingState.completed) {
-      return IconButton(
-        icon: const Icon(MdiIcons.pause),
-        iconSize: size.width * 0.1,
-        onPressed: pause,
-        splashColor: Colors.transparent,
-      );
-    } else {
-      return IconButton(
-        icon: const Icon(MdiIcons.replay),
-        iconSize: 64.0,
-        onPressed: () => audioPlayer!
-            .seek(Duration.zero, index: audioPlayer!.effectiveIndices!.first),
-      );
-    }
-  }
 }
